eliminar instruccion

+++ Simulacion/bandeja.h
@@ -6,13 +6,13 @@ struct Bandeja{
     int cantMax;
     int cantMin;
     int cantidad;
-    int velocidad;
+    double velocidad;
 
     Bandeja(){
         cantMax = 0;
         cantMin = 0;
         cantidad = 0;
-        velocidad = 0;
+        velocidad = 0.0;
     }
 };
